requirements.txt
fastapi
uvicorn
python-docx
pandas
openpyxl
jinja2